requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
atomicwrites==1.3.0
attrs==19.3.0
autobahn==19.11.1
autopep8==1.4.4
backcall==0.1.0
beautifulsoup4==4.7.1
bleach==3.1.0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
chillin-client==1.3.1
chillin-server==2.2.0
circuits==3.2
colorama==0.4.1
config==0.4.2
configparser==3.5.0
channels==3.0.4
cryptography==2.8
cycler==0.10.0
decorator==4.4.1
defusedxml==0.6.0
Django==2.2.5
django-filter==21.1
djangorestframework==3.12.4
entrypoints==0.3
enum34==1.1.6
Equation==1.2.1
html5lib==1.0.1
idna==2.7
importlib-metadata==0.23
ipykernel==5.1.3
ipython==7.9.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.15.1
Jinja2==2.10.3
jsonschema==3.1.1
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.0.0
jupyter-core==4.6.1
jupyter-server-proxy==1.1.0
kiwisolver==1.1.0
koala-serializer==0.8.0
MarkupSafe==1.1.1
matplotlib==3.1.1
mistune==0.8.4
more-itertools==7.2.0
mpmath==1.1.0
multidict==4.5.2
nbconvert==5.6.1
nbformat==4.4.0
notebook==6.0.2
numpy==1.17.2
opencv-contrib-python==4.5.4.60
opencv-python==4.5.4.60
opencv-python-headless==4.5.4.60
packaging==19.2
pandocfilters==1.4.2
parso==0.5.1
phonenumbers==8.10.2
pickleshare==0.7.5
pluggy==0.13.1
prometheus-client==0.7.1
prompt-toolkit==2.0.10
py==1.8.0
pycodestyle==2.5.0
pycparser==2.19
PyDbLite==3.0.4
Pygments==2.4.2
pyparsing==2.4.2
pyrsistent==0.15.5
pytest==5.3.0
python-dateutil==2.8.0
pytz==2019.2
pywin32==226
pywinpty==0.5.5
pyzmq==18.1.0
qtconsole==4.5.5
requests==2.20.0
selenium==3.141.0
Send2Trash==1.5.0
simpervisor==0.3
six==1.12.0
soupsieve==1.9.5
sqlparse==0.3.0
sympy==1.4
terminado==0.8.2
testpath==0.4.4
tornado==6.0.3
traitlets==4.3.3
txaio==18.8.1
urllib3==1.24.3
vpnotebook==0.1.3
vpython==7.5.2
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.5.1
yarl==1.3.0
youtube-dl==2019.11.28
zipp==0.6.0
